Add a `SqlBase` class from which `PpdbSql` inherits
This extracts generic SQL functionality from `PpdbSql` into a `SqlBase`
class. `PpdbSql` now inherits from this class.
