+++ python/lsst/dax/ppdb/_factory.py
@@ -49,7 +49,7 @@ def ppdb_type(config: PpdbConfig) -> type[PpdbSql]:
         Raised if type of ``config`` does not match any known types.
     """
     from .bigquery._config import PpdbBigQueryConfig
-    from .sql import PpdbSqlConfig
+    from .sql.config import PpdbSqlConfig
 
     if type(config) is PpdbSqlConfig:
         from .sql import PpdbSql
@@ -112,7 +112,7 @@ def config_type_for_name(type_name: str) -> type[PpdbConfig]:
         Raised if ``type_name`` does not match any known types.
     """
     if type_name == "sql":
-        from .sql import PpdbSqlConfig
+        from .sql.config import PpdbSqlConfig
 
         return PpdbSqlConfig
     elif type_name == "bigquery":
